requirements-vendor.txt
# these are libraries not bundled in standard appengine
# these need to be packaged with the app in lib folder using:
# pip install -r requirements-vendor.txt -t lib/
Django==1.10.3
django-filter==0.15.3
djangorestframework==3.5.2
Markdown==2.6.7